fix: elevator custom colors, economy int overflow, tpa accept scheduler error
- ElevatorManager: use FormatUtil.component() instead of Component.text(
  FormatUtil.color(...)) so &%<char> custom colors (which expand to the
  §x§R§R§G§G§B§B hex escape) are actually deserialized into a real hex
  Style instead of being sent as literal §-text, where each hex digit
  gets misread as its own standalone legacy color code.
- EconomyManager/PayCommand/EcoCommand: store and parse balances as long
  instead of int. The coins column is SQLite INTEGER (64-bit), but the
  Java side truncated through int/Integer.parseInt everywhere, capping
  balances at Integer.MAX_VALUE (~21.47億).
- TpaCore: pass 1L (not 0L) as the initial delay to the per-player
  countdown's runAtFixedRate(). Paper's EntityScheduler requires
  initialDelayTicks >= 1 and throws IllegalArgumentException otherwise;
  since this only runs from the accept path, every /tpaccept and
  /tphaccept with a non-zero teleport-delay-seconds errored.

+++ src/main/java/org/craftcore/stellaria/commands/EcoCommand.java
@@ -48,9 +48,9 @@ public class EcoCommand implements CommandExecutor, TabCompleter {
             return true;
         }
 
-        int amount;
+        long amount;
         try {
-            amount = Integer.parseInt(args[2]);
+            amount = Long.parseLong(args[2]);
             boolean valid = subCommand.equals("set") ? amount >= 0 : amount > 0;
             if (!valid) {
                 throw new NumberFormatException();
@@ -82,7 +82,7 @@ public class EcoCommand implements CommandExecutor, TabCompleter {
         return true;
     }
 
-    private void notify(CommandSender sender, OfflinePlayer target, String senderKey, String receiverKey, int amount) {
+    private void notify(CommandSender sender, OfflinePlayer target, String senderKey, String receiverKey, long amount) {
         String amountText = plugin.getEconomyManager().format(amount);
         sender.sendMessage(plugin.getConfigManager().getMessage(senderKey, target)
                 .replace("%amount%", amountText));
